Revert "Changed longs to ints"
This reverts commit c19ed407f7c9b0ed579ce84d3f9581b550788136.

+++ BottleType.h
@@ -7,7 +7,7 @@ class BottleType
 {
 public:
     String Name;
-    unsigned int MinWeight, MaxWeight, LiquidWeight;
+    unsigned long MinWeight, MaxWeight, LiquidWeight;
 
     BottleType();
     BottleType(String);
@@ -18,7 +18,7 @@ public:
 const BottleType UNKNOWN_BOTTLE PROGMEM = BottleType("unknown");
 const int MAX_BOTTLE_TYPES = 6;
 
-BottleType getBottleBasedOnWeight(int, BottleType[MAX_BOTTLE_TYPES]);
+BottleType getBottleBasedOnWeight(long, BottleType[MAX_BOTTLE_TYPES]);
 void loadBottles(BottleType bottleTypes[MAX_BOTTLE_TYPES]);
 
 #endif
